+++ src/main.ts
@@ -1,48 +1,26 @@
 import { setupProducts, getProductElement } from "./products";
+import { findElement } from "./utiles";
 import { setupCounter } from "./counter";
 import "./style.css";
-import { ICountMap } from "./types/mainType";
 import { setupCart } from "./cart";
-
-function findElement(
-  startingElement: HTMLElement,
-  selector: string
-): HTMLElement | null {
-  let currentElement: HTMLElement | null = startingElement;
-  while (currentElement) {
-    if (currentElement.matches(selector)) {
-      return currentElement;
-    }
-    currentElement = currentElement.parentElement;
-  }
-  return null;
-}
-
-// function sumAllCounts(countMap: ICountMap): number {
-//   let sum = 0;
-//   Object.values(countMap).forEach((number) => {
-//     sum += number;
-//   });
-//   return sum;
-
-//   // Alternative using reduce
-//   // return Object.values(countMap).reduce((total, current) => {
-//   //   total += current;
-//   //   return total;
-//   // }, 0);
-// }
 
 async function main(): Promise<void> {
   const { updateCount: updateProductCount, getProductById } =
     await setupProducts({
       container: document.querySelector("#products"),
+      onDecreaseClick: onDecreaseClick,
+      onIncreseClick: onIncreseClick,
     });
 
   const {
-    addProduct,
-    removeProduct,
+    addProduct: addProductToCart,
+    removeProduct: removeProductFromCart,
     updateCount: updateCartCount,
-  } = setupCart({ container: document.querySelector(".cart_items") });
+  } = setupCart({
+    container: document.querySelector(".cart_items"),
+    onDecreaseClick: onDecreaseClick,
+    onIncreseClick: onIncreseClick,
+  });
 
   // 이부분이 프로덕트에도 쓰이고 카트에서도 쓰이지 않을까 => 데코레이터 패턴으로 묶어서
   // const countMap: ICountMap = {};
@@ -54,29 +32,29 @@ async function main(): Promise<void> {
     totlaCountElement.innerHTML = totalCount.toString();
   };
 
-  const increaseCount = (productId: string): void => {
+  function onIncreseClick({ productId }: { productId: string }): void {
     const count = increase({ productId });
 
     updateProductCount({ productId, count });
     if (count === 1) {
-      addProduct({ product: getProductById({ productId }) });
+      addProductToCart({ product: getProductById({ productId }) });
     }
     updateCartCount({ productId, count });
     updateTotalCount(getTotalCount());
     // updateCart();
-  };
+  }
 
-  const decreaseCount = (productId: string): void => {
+  function onDecreaseClick({ productId }: { productId: string }): void {
     const count = decrease({ productId });
 
     updateProductCount({ productId, count });
     if (count === 0) {
-      removeProduct({ product: getProductById({ productId }) });
+      removeProductFromCart({ product: getProductById({ productId }) });
     }
     updateCartCount({ productId, count });
     updateTotalCount(getTotalCount());
     // updateCart();
-  };
+  }
 
   const productsElement = document.querySelector<HTMLElement>("#products");
 
@@ -91,35 +69,16 @@ async function main(): Promise<void> {
       targetElement.matches(".btn-increase")
     ) {
       if (targetElement.matches(".btn-decrease")) {
-        decreaseCount(productId);
+        onDecreaseClick({ productId });
+        // decreaseCount(productId);
       } else if (targetElement.matches(".btn-increase")) {
-        increaseCount(productId);
+        onIncreseClick({ productId });
+        // increaseCount(productId);
       }
     }
   });
 
   const cartItemsElement = document.querySelector<HTMLElement>(".cart_items");
-
-  cartItemsElement?.addEventListener("click", (event) => {
-    console.log("aaa", event.target);
-    const targetElement = event.target as HTMLElement;
-    const productElement = findElement(targetElement, ".product")!;
-
-    const productId = productElement.getAttribute("data-product-id")!;
-    console.log("p", productId);
-
-    if (
-      targetElement.matches(".btn-decrease") ||
-      targetElement.matches(".btn-increase")
-    ) {
-      if (targetElement.matches(".btn-decrease")) {
-        decreaseCount(productId);
-      } else if (targetElement.matches(".btn-increase")) {
-        console.log("hello");
-        increaseCount(productId);
-      }
-    }
-  });
 
   const btnCartElement = document.querySelector<HTMLElement>(".btn-cart");
